Refactor common decorators and response models, removing unused imports and consolidating route definitions in the timings module
